requirements.txt
atomicwrites==1.4.0
attrs==21.2.0
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
colorama==0.4.4
Django==3.0.5
django-filter==2.4.0
djangorestframework==3.11.0
idna==2.10
importlib-metadata==4.3.1
iniconfig==1.1.1
packaging==20.9
pep517==0.10.0
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.4
pytest-django==4.3.0
requests==2.25.1
sqlparse==0.3.1
toml==0.10.2
typing-extensions==3.10.0.0
urllib3==1.26.5
zipp==3.4.1
asgiref==3.2.10
djangorestframework-simplejwt==4.3.0
gunicorn==20.0.4
psycopg2-binary==2.8.5
PyJWT==1.7.1
pytz==2020.1